api/requirements.txt
fastapi
uvicorn
pandas
google-cloud-bigquery
python-jose
firebase-admin
stripe
sentence-transformers
faiss-cpu
